extract shared test utilities to avoid duplication
